database-scheduler/database-scheduler/requirements.txt
cassandra-driver==3.25.0
mysql-connector-python==8.0.29